organize fetch types
